fix(extraction): resolve mypy errors in spatial_data and extraction_request
- spatial_data: wrap confidence returns with float() to fix no-any-return
- extraction_request: replace isinstance(EntityId) NewType check with str guard
- extraction_request: add return type annotations to source_id and document_id properties
- extraction_request: guard language property against None with fallback to empty string

+++ processing/extraction/spatial_data.py
@@ -296,15 +296,15 @@ class ExtractableSpan:
 
 def _word_based_confidence(line: Any) -> ConfidenceScore:
     if not line.has_words:
-        return line.confidence
-    return round(sum(w.confidence for w in line.words) / len(line.words), 4)
+        return float(line.confidence)
+    return float(round(sum(w.confidence for w in line.words) / len(line.words), 4))
 
 
 def _block_confidence_from_words(block: Any) -> ConfidenceScore:
     words = [w for line in block.lines for w in line.words]
     if not words:
-        return block.confidence
-    return round(sum(w.confidence for w in words) / len(words), 4)
+        return float(block.confidence)
+    return float(round(sum(w.confidence for w in words) / len(words), 4))
 
 
 def _build_block_spans(page: Any) -> list[ExtractableSpan]:
